群聊天界面布局实现和简单群成员加载显示

+++ App/iTalker/factory/src/main/java/com/imist/italker/factory/presenter/message/ChatContact.java
@@ -38,7 +38,7 @@ public interface ChatContact {
         //是否是管理员
         void showAdminOption(boolean isAdmin);
         //初始化成员
-        void onInitGroupMembers(List<MemberUserModel> members ,int moreCount );
+        void onInitGroupMembers(List<MemberUserModel> members ,long moreCount );
 
 
     }
